+++ parser.c
@@ -1378,7 +1378,7 @@ int readitem1(int th)
 	return (makeflt(atof(stok.buf)));
 	case LONGNUM:
 	char *longptr;
-	long longnum;
+	long long int longnum;
 	longnum = strtol(stok.buf,&longptr,10);
 	return (makelong(longnum));
     case BIGNUM:
